feat(api): rename BuildSpec as Pipeline spec
What we are calling it a Build it can be better thought as a Pipeline. For this reason it makes sense renaming (at least, the spec, mostly used internally).
The Build type could still stay as externally the user sees it as an action required to build the Camel application.

Ref #4024